refactor to use hook
